Congressional app challenge commit
This is the working version of the code that was run locally for the recording of the demo. Commited to github on 10/29, before the submission date. Please keep in mind that we will continue to update this github because we are still bettering our application, and a commit might also be made on 10/30 before the deadline